source: move files to Displayable folder
