+++ Informant/Implementation/Displayable/NewRecipeDisplayable.cs
@@ -6,7 +6,7 @@ using StardewValley.Menus;
 using Color = Microsoft.Xna.Framework.Color;
 using Rectangle = Microsoft.Xna.Framework.Rectangle;
 
-namespace Slothsoft.Informant.Implementation;
+namespace Slothsoft.Informant.Implementation.Displayable;
 
 internal class NewRecipeDisplayable : IDisplayable
 {
@@ -75,8 +75,8 @@ internal class NewRecipeDisplayable : IDisplayable
 
         // recipe.timesCrafted is not updated it seems
         var timesCrafted = recipe.isCookingRecipe
-                ? (Game1.player.recipesCooked.ContainsKey(recipe.getIndexOfMenuView()) ? Game1.player.recipesCooked[recipe.getIndexOfMenuView()] : 0)
-                : (Game1.player.craftingRecipes.ContainsKey(recipe.name) ? Game1.player.craftingRecipes[recipe.name] : 0);
+                ? Game1.player.recipesCooked.ContainsKey(recipe.getIndexOfMenuView()) ? Game1.player.recipesCooked[recipe.getIndexOfMenuView()] : 0
+                : Game1.player.craftingRecipes.ContainsKey(recipe.name) ? Game1.player.craftingRecipes[recipe.name] : 0;
         if (timesCrafted > 0) {
             // we are on the recipe page, have a recipe which was already craftet? nice, it's not new
             return;
